pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "resontinex"
version = "2.1.0"
description = "Production-grade AI workflow orchestration system with intelligent overlay routing, circuit breaker protection, and comprehensive drift detection"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "RESONTINEX Team", email = "chris@custaa.com"}
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: System :: Distributed Computing",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
keywords = ["ai", "workflow", "orchestration", "governance", "monitoring"]

# Minimal core dependencies for light installation
dependencies = [
    "pydantic>=1.10.0,<3.0.0",
    "typing-extensions>=4.0.0; python_version<'3.10'",
]

[project.optional-dependencies]
# Light installation - core functionality only
light = []

# Full installation with all features
full = [
    "prometheus-client>=0.16.0",
    "rich>=13.0.0",
    "typer>=0.9.0",
    "pyyaml>=6.0.0",
    "requests>=2.28.0",
    "packaging>=21.0",
]

# Development dependencies
dev = [
    "pre-commit>=3.0.0",
    "ruff>=0.1.6",
    "black>=23.11.0",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "bandit>=1.7.5",
    "isort>=5.12.0",
    "yamllint>=1.33.0",
    "mypy>=1.8.0",
    "types-PyYAML>=6.0.0",
    "types-requests>=2.28.0",
]

# Documentation dependencies
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.24.0",
    "pdoc>=14.4.0",
]

# Architecture analysis tools
arch = [
    "pydeps>=1.12.0",
    "pyan3>=1.2.0",
]

# All dependencies combined
all = [
    "resontinex[full,docs,arch]",
]

[project.urls]
Homepage = "https://resontinex.dev"
Documentation = "https://docs.resontinex.dev"
Repository = "https://github.com/resontinex/resontinex"
"Bug Tracker" = "https://github.com/resontinex/resontinex/issues"

[project.scripts]
resontinex = "scripts.rnx:app"
rnx = "scripts.rnx:app"

[project.entry-points."resontinex.plugins"]
governance = "resontinex_governance:init"

[tool.setuptools.packages.find]
include = ["resontinex*", "resontinex_governance*"]
exclude = ["tests*", "test_results*", "docs*", "examples*"]

[tool.setuptools.package-data]
resontinex = ["py.typed"]

[tool.ruff]
# Ruff configuration for code quality
target-version = "py38"
line-length = 120
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "S",   # bandit
]
ignore = [
    "E501",  # line too long, handled by black
    "S101",  # use of assert
    "S608",  # hardcoded SQL query
    "ARG002", # unused method argument
    "ARG001", # unused function argument
]

# Exclude directories
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

[tool.ruff.per-file-ignores]
"tests/*" = ["S101", "ARG", "S106"]  # Allow asserts and test-specific patterns
"scripts/*" = ["S108", "S105"]      # Allow temp files and hardcoded passwords in scripts

[tool.black]
line-length = 120
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 120
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--cov=resontinex",
    "--cov-report=term-missing",
    "--cov-report=html:test_results/coverage",
    "--cov-fail-under=80",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.bandit]
exclude_dirs = ["tests", "test_results"]
skips = ["B101", "B601", "B608"]  # Skip assert, shell injection (controlled usage), SQL injection (controlled usage)

[tool.coverage.run]
source = ["resontinex", "scripts"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__init__.py",
    "*/conftest.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.mypy]
# Strict mode for fusion_ops module
python_version = "3.8"
strict = true
files = ["fusion_ops/**/*.py"]
ignore_missing_imports = true
show_error_codes = true
warn_return_any = true
warn_unused_configs = true
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

# Incremental adoption - start strict on fusion_ops, relax for others
[[tool.mypy.overrides]]
module = [
    "resontinex.*",
    "scripts.*",
    "tests.*"
]
ignore_missing_imports = true
check_untyped_defs = false
disallow_untyped_defs = false
disallow_incomplete_defs = false